app/llm/gemini_provider.py
"""
Google Gemini provider for external LLM generation.

Provides integration with Google's Gemini API for text generation
with proper consent and privacy handling.
"""

from typing import List, Dict, Any, Optional
import google.generativeai as genai

from .base import BaseLLMProvider
from ..logging_conf import get_logger

logger = get_logger(__name__)


class GeminiProvider(BaseLLMProvider):
    """Google Gemini provider for external text generation."""
    
    def __init__(
        self, 
        api_key: str, 
        model: str = "gemini-2.5-flash",
        safety_settings: Optional[Dict[str, Any]] = None
    ):
        super().__init__(model)
        self.api_key = api_key
        self.safety_settings = safety_settings or self._default_safety_settings()
        
        # Configure the API
        if api_key:
            genai.configure(api_key=api_key)
        
        self.model = None
        if api_key:
            try:
                self.model = genai.GenerativeModel(
                    model_name=model,
                    safety_settings=self.safety_settings
                )
            except Exception as e:
                logger.error("Failed to initialize Gemini model", error=str(e))
    
    async def generate(
        self,
        system: str,
        messages: List[Dict[str, str]],
        max_tokens: int = 900,
        temperature: float = 0.2
    ) -> str:
        """
        Generate text using Google Gemini API.
        
        Args:
            system: System prompt
            messages: Conversation messages  
            max_tokens: Maximum tokens to generate
            temperature: Generation temperature
            
        Returns:
            Generated text response
        """
        # TODO: Implement Gemini generation
        raise NotImplementedError("Gemini generation not yet implemented")
    
    async def test_connection(self) -> bool:
        """Test connection to Gemini API."""
        if not self.api_key or not self.model:
            return False
        
        try:
            # Simple test generation
            response = self.model.generate_content(
                "Test connection. Respond with 'OK'.",
                generation_config=genai.types.GenerationConfig(
                    max_output_tokens=10,
                    temperature=0.1,
                )
            )
            return "OK" in response.text.upper()
        except Exception as e:
            logger.error("Gemini connection test failed", error=str(e))
            return False
    
    def _default_safety_settings(self) -> List[Dict[str, Any]]:
        """Get default safety settings for Gemini."""
        return [
            {
                "category": "HARM_CATEGORY_HARASSMENT",
                "threshold": "BLOCK_MEDIUM_AND_ABOVE"
            },
            {
                "category": "HARM_CATEGORY_HATE_SPEECH", 
                "threshold": "BLOCK_MEDIUM_AND_ABOVE"
            },
            {
                "category": "HARM_CATEGORY_SEXUALLY_EXPLICIT",
                "threshold": "BLOCK_MEDIUM_AND_ABOVE"
            },
            {
                "category": "HARM_CATEGORY_DANGEROUS_CONTENT",
                "threshold": "BLOCK_MEDIUM_AND_ABOVE"
            },
        ]
    
    def _format_messages_for_gemini(
        self, 
        system: str, 
        messages: List[Dict[str, str]]
    ) -> str:
        """Format messages into single prompt for Gemini."""
        # TODO: Implement message formatting for Gemini
        raise NotImplementedError("Message formatting not yet implemented")