requirements.txt
# Core application framework
nicegui>=1.4.0

# Vector database and embeddings
chromadb>=0.4.0

# PDF processing
pymupdf>=1.23.0
ocrmypdf>=15.0.0

# Data validation and models
pydantic>=2.0.0

# Web framework and async support
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Structured logging
structlog>=23.0.0

# LLM providers
ollama>=0.2.0
google-generativeai>=0.5.0

# Agent memory system
letta>=0.3.0

# Configuration management  
tomli>=2.0.0

# HTTP client
aiohttp>=3.8.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Type checking (development)
mypy>=1.0.0

# Optional: Code formatting (development)
black>=23.0.0

# Optional: Additional utilities
python-dotenv>=1.0.0